requirements.txt
absl-py==2.1.0
accelerate==0.21.0
aiosignal==1.3.1
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
boltons==23.1.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudinary==1.38.0
colorama==0.4.6
coloredlogs==15.0.1
einops==0.4.1
exceptiongroup==1.2.0
fastapi==0.111.0
filelock==3.13.1
flatbuffers==23.5.26
frozenlist==1.4.1
fsspec==2024.2.0
gast==0.5.4
google-pasta==0.2.0
grpcio==1.62.1
h11==0.14.0
h5py==3.10.0
httptools==0.6.1
huggingface-hub==0.20.3
humanfriendly==10.0
idna==3.6
imageio==2.34.0
Jinja2==3.1.3
keras==3.1.1
lazy_loader==0.4
libclang==18.1.1
lightning-utilities==0.10.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
multidict==6.0.5
namex==0.0.7
networkx==3.2.1
numpy==1.25.2
omegaconf==2.2.3
onnxruntime==1.16.3
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
protobuf==4.25.2
psutil==5.9.5
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
Pygments==2.17.2
pyreadline3==3.4.1
python-dotenv==1.0.1
PyYAML==6.0
regex==2023.12.25
requests==2.31.0
rich==13.7.1
safetensors==0.3.1
scikit-image==0.23.2
scipy==1.13.1
six==1.16.0
sniffio==1.3.0
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
tifffile==2024.2.12
tokenizers==0.13.3
torch==2.2.0
torchmetrics==1.3.0.post0
tqdm==4.64.1
trampoline==0.1.2
typing_extensions==4.9.0
urllib3==2.2.0
uvicorn==0.23.2
watchfiles==0.21.0
websockets==12.0
Werkzeug==3.0.2
wrapt==1.16.0
yarl==1.9.4
supabase~=2.5.0
Pillow==9.2.0